+++ doochul/src/main/java/org/doochul/infra/KakaoLoginUserClient.java
@@ -1,4 +1,4 @@
-package org.doochul.domain.oauth.kakao;
+package org.doochul.infra;
 
 import org.doochul.ui.dto.KakaoUserInfoResponse;
 import org.springframework.stereotype.Component;
@@ -6,12 +6,12 @@ import org.springframework.web.reactive.function.client.WebClient;
 import reactor.core.publisher.Flux;
 
 @Component
-public class KakaoUserInfo {
+public class KakaoLoginUserClient {
 
     private final WebClient webClient;
     private static final String USER_INFO_URI = "https://kapi.kakao.com/v2/user/me";
 
-    public KakaoUserInfo() {
+    public KakaoLoginUserClient() {
         this.webClient = WebClient.create(USER_INFO_URI);
     }
